création des components FileHandling et Footer
