+++ components/pending.js
@@ -1,4 +1,3 @@
-import { render } from 'react-dom'
 import styles from '../styles/Home.module.css'
 
 export default function PendingMessage({
