+++ pages/lecteur/cnaf/beneficiaire/index.js
@@ -4,8 +4,8 @@ import Head from 'next/head'
 import Admin from '../../../../components/admin'
 import ResponsiveCalendar from '../../../../components/chart'
 import Layout from '../../../../components/layout'
-import Mailer from '../../../../components/mailer'
-import PendingMessage from '../../../../components/pending'
+import FileHandler from '../../../../components/file'
+import Footer from '../../../../components/footer'
 import styles from '../../../../styles/Home.module.css'
 
 import { frequencyNames, typeNames } from '../../../../lib/cnaf'
@@ -38,13 +38,6 @@ export default function Beneficiaire() {
       fileHandler(devFile)
     }
   }, [devFile])
-
-  const selectHandler = useCallback((event) => {
-    for (var i = 0; i<event.target.files.length; i++) {
-      fileHandler(event.target.files[i])
-    }
-    event.target.value = ''
-  })
 
   const handleDateHistogram = useCallback(event => showDateHistogram(parseInt(event.target.dataset.index)))
   const showDateHistogram = useCallback(index => {
@@ -140,21 +133,7 @@ export default function Beneficiaire() {
           Lecteur de fichier<br/>« Bénéficiaire » de la CNAF
         </h1>
 
-        <p className={styles.description}>
-          Glissez et déposez le fichier CNAF à analyser ou sélectionnez le.<br/>
-          <input type="file" onChange={selectHandler} multiple/>
-        </p>
-
-        {isPending && <PendingMessage fileSize={fileSize}/>}
-
-        <p className={styles.description}>
-          Les opérations sont toutes réalisées sur votre ordinateur.<br/>
-          Aucune donnée personnelle n'est transférée.
-        </p>
-
-        <p className={styles.description}>
-          <a href="#pourquoi">Pourquoi un tel lecteur&nbsp;?</a>
-        </p>
+        <FileHandler fileHandler={fileHandler} isPending={isPending} fileSize={fileSize} />
 
         { runs && runs.length > 0 && (<>
           <h2 className={styles.subtitle}>
@@ -213,21 +192,14 @@ export default function Beneficiaire() {
           </>
         )}
 
-        <p className={styles.description}>
-          Un problème, une question ? Contactez-nous à <Mailer subject="Flux bénéficiaire CNAF">data.insertion@beta.gouv.fr</Mailer>
-        </p>
-
-        <h2 id="pourquoi" className={styles.subtitle}>
-          Pourquoi un lecteur de fichier CNAF&nbsp;?
-        </h2>
-
-        <p className={styles.text}>
-          Tous les départements n'ont pas les outils pour analyser les fichiers envoyés par la CNAF. Cela peut ralentir et compliquer nos échanges.
-        </p>
-        <p className={styles.text}>
-          Avec cet outil, nous souhaitons permettre aux personnes qui ont accès à ces fichiers d'en extraire des statistiques générales sans avoir à mettre les mains dans le camboui elles-même.
-        </p>
-
+        <Footer
+          subject="Flux bénéficiaire CNAF"
+          pourquoi="Pourquoi un lecteur de fichier CNAF&nbsp;?"
+          text={<>
+            <p className={styles.text}>Tous les départements n'ont pas les outils pour analyser les fichiers envoyés par la CNAF. Cela peut ralentir et compliquer nos échanges.</p>
+            <p className={styles.text}>Avec cet outil, nous souhaitons permettre aux personnes qui ont accès à ces fichiers d'en extraire des statistiques générales sans avoir à mettre les mains dans le camboui elles-même.</p>
+          </>}
+        />
       </main>
     </Layout>
   )
